# dot-connect/docs/swagger.yaml
definitions:
  model.PostReport:
    properties:
      content:
        type: string
      location:
        example: ex_location
        type: string
      locationDetail:
        example: ex_location_detail
        type: string
    type: object
  model.Report:
    properties:
      content:
        type: string
      id:
        example: 1
        type: integer
      image:
        type: string
      location:
        example: ex_location
        type: string
    type: object
info:
  contact: {}
paths:
  /reports:
    get:
      description: get MyReports
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Report'
      summary: List reports
      tags:
      - reports
  /reports/{reportId}:
    get:
      consumes:
      - application/json
      description: get one of report
      parameters:
      - description: report ID
        in: path
        name: reportId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Report'
      summary: get report detail
      tags:
      - reports
  /reports/upload:
    post:
      description: 에러 신고에 이미지, 장소, 내용이 담긴다.
      parameters:
      - in: formData
        name: content
        type: string
      - example: ex_location
        in: formData
        name: location
        type: string
      - example: ex_location_detail
        in: formData
        name: locationDetail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PostReport'
      summary: 에러 신고 업로드
      tags:
      - reports
  /translation:
    get:
      description: 한글를 점자로 점역하는 API
      produces:
      - application/json
      responses: {}
      summary: 한글 -> 점자 점역
      tags:
      - braille
  /translation/braille:
    post:
      description: 점자를 점자로 역점역하는 API
      produces:
      - application/json
      responses: {}
      summary: 점자 -> 한글 역점역
      tags:
      - braille
swagger: "2.0"